requirements.txt
opencv-python
numpy
torch
pickle-mixin
matplotlib-pyplot
PIL  
streamlit_shadcn_ui 
